pyproject.toml
[project]
name = "pdf-production-rag"
version = "0.1.0"
description = "Basic RAG app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "inngest>=0.5.12",
    "llama-index-core>=0.14.8",
    "llama-index-readers-file>=0.5.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.0",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
